Move dump_hex() to debug.py
